Update script1.py
Removed functions for right and left click and replaced with one function for all clicks.
Realised the bank window can be opened directly by left clicking on the desk in front of the banker. So the enterbank() function has been completely changed to reflect that and is now 1 line long.